Move presentationals out of main component file
